+++ packages/hippo_core/hippo_core_flutter/lib/src/animation.dart
@@ -7,12 +7,4 @@
 // SPDX-License-Identifier: LicenseRef-Hipposphere-Proprietary
 // ---------------------------------------------------------------------------
 */
-export 'src/animation.dart';
-export 'src/bloc.dart';
-export 'src/common.dart';
-export 'src/data_subject/data_subject.dart';
-export 'src/filter.dart';
-export 'src/layout.dart';
-export 'src/resource.dart';
-export 'src/sliver.dart';
-export 'src/store/store.dart';
+export 'animation/frame_rate_ticker_provider.dart';